training/requirements.txt
torch==2.2.0
torchvision==0.17.0
Pillow==10.1.0
tqdm==4.66.1
numpy>=1.24.3
